Improve compatibility with Async Particles
Async Particles appears to call SingleQuadParticleMixin.render in multiple threads, which, when combined with this mod, causes a race condition as Angled Particles uses a global particle rotation variable to conserve memory. This patch moves that global variable into a thread-local, which appears to fully solve the race condition and fix compatibility in my (limited) testing.